+++ CHANGELOG.rst
@@ -6,9 +6,9 @@ HEAD (2020-10-??)
 
 *	Added the following attributes to ``AppLookupControl``:
 
-	-	``local_master_control``
-	-	``local_detail_controls``
-	-	``remote_master_control``
+	-	``local_master_control``,
+	-	``local_detail_controls``,
+	-	``remote_master_control``.
 
 
 0.13.0 (2020-09-17)